+++ db/migrate/20170110184430_add_rating_to_datasets.rb
@@ -1,0 +1,6 @@
+class AddRatingToDatasets < ActiveRecord::Migration[5.0]
+  def change
+    add_column :datasets, :rating, :float
+    add_index :datasets, :rating
+  end
+end