+++ db/migrate/20170110185244_create_dataset_reviews.rb
@@ -1,0 +1,12 @@
+class CreateDatasetReviews < ActiveRecord::Migration[5.0]
+  def change
+    create_table :dataset_reviews do |t|
+      t.integer :dataset_id
+      t.integer :user_id
+      t.integer :rating
+      t.text :review
+      t.index [:dataset_id, :user_id], unique: true
+      t.timestamps
+    end
+  end
+end