+++ db/migrate/20170110190802_add_dataset_review_id_to_notifications.rb
@@ -1,0 +1,8 @@
+class AddDatasetReviewIdToNotifications < ActiveRecord::Migration[5.0]
+  def change
+    add_column :notifications, :dataset_id, :integer
+    add_index :notifications, :dataset_id
+    add_column :notifications, :dataset_review_id, :integer
+    add_index :notifications, :dataset_review_id
+  end
+end