+++ test/controllers/community_tools_controller_test.rb
@@ -2,7 +2,7 @@
 
 require 'test_helper'
 
-# Test for admin side of community tools
+# Test for admin side of community tools.
 class CommunityToolsControllerTest < ActionController::TestCase
   setup do
     @community_tool = community_tools(:published)